wasm: move typescript definitions to their own file
